// shared/enum/features.enum.ts
export enum FeatureName {
    DINING_ROOM = 'Dinning room',
    DELIVERY = 'Delivery',
    HIGH_TABELS = 'High tabels',
    STANDING_TABELS = 'Standing tabels',
    POOL_TABELS = 'Pool tabels',
    PRIVATE_TABELS = 'Private tabels',
    BEACH_TABELS = 'Beach tabels',
    FOOD_PICK_UP = 'Food pickup',
    TERRACE = 'Terrace',
    BAR = 'Bar',
    BEACH_BAR = 'Beach bar',
    ROOF_TOP = 'Roof top',
    BACK_GARDEN = 'Back garden',
    LACK_VIEW = 'Lack view',
    RIVER_VIEW = 'River view',
    PANORAMA_VIEW = 'Panorama view',
    SEA_VIEW = 'Sea view',
    POOL_VIEW = 'Pool view',
    MOUNTAIN_VIEW = 'Mountain view',
    CITY_CENTER = 'City center',
    CATERING = 'Catering',
    DANCE_FLOOR = 'Dance floor',
    LIVE_MUSIC = 'Live Music',
    LIVE_DJ = 'Live dj',
    LIVE_BAND = 'Live band',
    LIVE_SHOW = 'Live show',
    GARDEN = 'Garden',
    KIDS_AREA = 'Kids area',
    DRIVE_THROUGH = 'Drive through',
    SAND_AREA = 'Sand area',
    GAMING_AREA = 'Gaming area',
    PARKING_ASSISTANCE = 'Parking assistance',
    PARKING_AREA = 'Parking area',
    UNDER_WATER = 'Under water',
    ELEVATOR = 'Elevator',
    POOL = 'Pool',
    ANIMAL_FRIENDLY = 'Animal friendly',
    KIDS_FRIENDLY = 'Kids friendly',
    WHEEL_CHAIR_FRIENDLY = 'Wheel chair friendly',
    BLIND_PEOPLE_FRIENDLY = 'Blind people friendly',
    SENIORS_FRIENDLY = 'Seniors friendly',
    UNDEFINED_FEATURE = 'UNDEFINED_FEATURE',
}